added part of car rental lld

+++ src/main/java/problems/movieticketbookingsystem/models/CashPaymentStrategy.java
@@ -1,8 +1,10 @@
 package problems.movieticketbookingsystem.models;
 
+import java.math.BigDecimal;
+
 public class CashPaymentStrategy implements IPaymentStrategy {
 
-    public void pay(int amount) {
+    public void pay(BigDecimal amount) {
         System.out.println("Payment of amount : " + amount + " is made by cash.");
     }
 }
